+++ spec/rubocop/cop/metrics/line_length_spec.rb
@@ -188,4 +188,102 @@ describe RuboCop::Cop::Metrics::LineLength, :config do
       end
     end
   end
+
+  context 'when IgnoreCopDirectives is disabled' do
+    let(:cop_config) { { 'Max' => 80, 'IgnoreCopDirectives' => false } }
+
+    context 'and the source is acceptable length' do
+      let(:acceptable_source) { 'a' * 80 }
+
+      context 'with a trailing Rubocop directive' do
+        let(:cop_directive) { ' # rubcop:disable Metrics/SomeCop' }
+        let(:source) { acceptable_source + cop_directive }
+
+        it 'registers an offense for the line' do
+          inspect_source(cop, source)
+          expect(cop.offenses.size).to eq(1)
+        end
+
+        it 'highlights the excess directive' do
+          inspect_source(cop, source)
+          expect(cop.highlights).to eq([cop_directive])
+        end
+      end
+
+      context 'with an inline comment' do
+        let(:excess_comment) { ' ###' }
+        let(:source) { acceptable_source + excess_comment }
+
+        it 'highlights the excess comment' do
+          inspect_source(cop, source)
+          expect(cop.highlights).to eq([excess_comment])
+        end
+      end
+    end
+
+    context 'and the source is too long and has a trailing cop directive' do
+      let(:excess_with_directive) { 'b # rubocop:disable Metrics/AbcSize' }
+      let(:source) { 'a' * 80 + excess_with_directive }
+
+      it 'highlights the excess source and cop directive' do
+        inspect_source(cop, source)
+        expect(cop.highlights).to eq([excess_with_directive])
+      end
+    end
+  end
+
+  context 'when IgnoreCopDirectives is enabled' do
+    let(:cop_config) { { 'Max' => 80, 'IgnoreCopDirectives' => true } }
+
+    context 'and the Rubocop directive is excessively long' do
+      let(:source) { <<-END }
+        # rubocop:disable Metrics/SomeReallyLongMetricNameThatShouldBeMuchShorterAndNeedsANameChange
+      END
+
+      it 'accepts the line' do
+        inspect_source(cop, source)
+        expect(cop.offenses).to be_empty
+      end
+    end
+
+    context 'and the Rubocop directive causes an excessive line length' do
+      let(:source) { <<-END }
+        def method_definition_that_is_just_under_the_line_length_limit(foo, bar) # rubocop:disable Metrics/AbcSize
+          # complex method
+        end
+      END
+
+      it 'accepts the line' do
+        inspect_source(cop, source)
+        expect(cop.offenses).to be_empty
+      end
+
+      context 'and has explanatory text' do
+        let(:source) { <<-END }
+          def method_definition_that_is_just_under_the_line_length_limit(foo) # rubocop:disable Metrics/AbcSize inherently complex!
+            # complex
+          end
+        END
+
+        it 'accepts the line' do
+          inspect_source(cop, source)
+          expect(cop.offenses).to be_empty
+        end
+      end
+    end
+
+    context 'and the source is too long' do
+      let(:source) { 'a' * 80 + 'bcd' + ' # rubocop:enable Style/ClassVars' }
+
+      it 'registers an offense for the line' do
+        inspect_source(cop, source)
+        expect(cop.offenses.size).to eq(1)
+      end
+
+      it 'highlights only the non-directive part' do
+        inspect_source(cop, source)
+        expect(cop.highlights).to eq(['bcd'])
+      end
+    end
+  end
 end