+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * [#3566](https://github.com/bbatsov/rubocop/issues/3566): Add new `Metric/BlockLength` cop to ensure blocks don't get too long. ([@savef][])
 * [#3428](https://github.com/bbatsov/rubocop/issues/3428): Add support for configuring `Style/PreferredHashMethods` with either `short` or `verbose` style method names. ([@abrom][])
 * [#3455](https://github.com/bbatsov/rubocop/issues/3455): Add new `Rails/DynamicFindBy` cop. ([@pocke][])
+* [#3542](https://github.com/bbatsov/rubocop/issues/3542): Add a configuration option, `IgnoreCopDirectives`, to `Metrics/LineLength` to stop cop directives (`# rubocop:disable Metrics/AbcSize`) from being counted when considering line length. ([@jmks][])
 
 ### Bug fixes
 
@@ -2410,3 +2411,4 @@
 [@logicminds]: https://github.com/logicminds
 [@abrom]: https://github.com/abrom
 [@thegedge]: https://github.com/thegedge
+[@jmks]: https://github.com/jmks/